Reorganize integration test suite
