+++ crates/shelterwood/tests/acceptance.rs
@@ -1,0 +1,8 @@
+//! Application-scale acceptance scenarios from `SPEC.md` Appendix C.
+
+#[path = "acceptance/assistant.rs"]
+mod assistant;
+#[path = "acceptance/shard_store.rs"]
+mod shard_store;
+#[path = "acceptance/sidecar.rs"]
+mod sidecar;
