+++ crates/shelterwood/tests/common/policy.rs
@@ -1,0 +1,5 @@
+use shelterwood::{Backoff, RestartCondition, RestartPolicy};
+
+pub fn never() -> RestartPolicy {
+    RestartPolicy::new(RestartCondition::Never, Backoff::Immediate)
+}
