+++ crates/shelterwood/tests/delivery.rs
@@ -1,11 +1,9 @@
 use std::{
-    future::Future,
-    pin::Pin,
     sync::{
         Arc, Mutex,
         atomic::{AtomicUsize, Ordering},
     },
-    task::{Context, Poll, Waker},
+    task::Poll,
     time::Duration,
 };
 
@@ -13,12 +11,7 @@ use shelterwood::{
     CallErrorKind, ExitError, ExitResult, Mailbox, RawActor, RawContext, RawDef, RawOnceDef, Reply,
     SendErrorKind, Shutdown, SubtreeOnceDef, Tree,
 };
-use shelterwood_test_support::{ReleaseGate, assert_quiet, poll_until};
-
-fn poll_once<F: Future>(future: Pin<&mut F>) -> Poll<F::Output> {
-    let mut context = Context::from_waker(Waker::noop());
-    future.poll(&mut context)
-}
+use shelterwood_test_support::{ReleaseGate, assert_quiet, poll_once, poll_until};
 
 struct PrefixActor {
     generation: usize,
