+++ crates/shelterwood/tests/readiness.rs
@@ -7,14 +7,15 @@ use std::{
 };
 
 use shelterwood::{
-    Backoff, DynamicTree, ExitError, ExitKind, Readiness, ReadinessDeadline, RestartCondition,
-    RestartPolicy, Shutdown, StartupError, StartupFailureCause, SubtreeOnceDef, TaskDef, Tree,
+    DynamicTree, ExitError, ExitKind, Readiness, ReadinessDeadline, Shutdown, StartupError,
+    StartupFailureCause, SubtreeOnceDef, TaskDef, Tree,
 };
 use shelterwood_test_support::{ReleaseGate, advance_time, assert_quiet, poll_until};
 
-fn never() -> RestartPolicy {
-    RestartPolicy::new(RestartCondition::Never, Backoff::Immediate)
-}
+#[path = "common/policy.rs"]
+mod policy;
+
+use policy::never;
 
 #[tokio::test]
 async fn ordered_startup_waits_for_manual_readiness() {
